[v0.91.2][maintainability] Split adl/src/acc.rs by responsibility (Closes #3155)
